tests passed for error and skip
The api now performs as expected regarding the triggering of events,
and basic promise callbacks (including multiples).

The next stage will be to refractor the state api and trigger methods
for clarity, reimplement automatic detection / triggering, and allow
grouping by having/installing queues as event listeners, which can
be located using a queued objects state.

It would also be prudent to note that I need a better method for
counting asynchronous operations.  The current rig doesn't allow
verifying the data at the end.

Signed-off-by: James Arlow <james.arlow@gmail.com>